+++ container-engine-lib/docs/changelog.md
@@ -1,6 +1,8 @@
 # TBD
 ### Features
+* The enclave data volume gets mounted on all services
 * `DockerKurtosisBackend.CreateEnclave` also creates an enclave data volume
+
 
 # 0.15.0
 ### Fixes